% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as_matrix}
\alias{as_matrix}
\title{Convert to matrix.}
\usage{
as_matrix(x, nrows, ncols)
}
\arguments{
\item{x}{Object to cast as matrix.}

\item{nrows}{Number of rows.}

\item{ncols}{Number of columns.}
}
\description{
Convert to matrix.
}

